pyproject.toml
[tool.poetry]
name = "datus-metricflow"
version = "0.1.0"
description = "Simplified MetricFlow integration for Datus Agent - One-command setup for semantic metrics"
authors = ["Datus.ai <support@datus.ai>", "Transform <hello@transformdata.io>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
homepage = "https://datus.ai"
repository = "https://github.com/Datus-ai/metricflow"
keywords = ["metrics", "sql", "analytics", "datus", "metricflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

packages = [
    {include = "metricflow"},
    {include = "datus_metricflow"}
]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
typing_extensions = ">=4.0.0"

croniter = "^1.3.4"
pycron = "^3.0.0"
pydantic = "^2.0.0"
fuzzywuzzy = "^0.18.0"
jsonschema = "3.2.0"
SQLAlchemy = "^1.4.42"
snowflake-sqlalchemy = "^1.4.3"
sqlalchemy-redshift = "0.8.1"
numpy = ">=1.22.2,<2.0"
pandas = "^1.3.0"
Jinja2 = ">=2.11.3"
PyYAML = "^6.0"
snowflake-connector-python = ">=2.7.8"
mo-sql-parsing = "^9.328.23003"
tabulate = "0.8.9"
more-itertools = "8.10.0"
graphviz = "0.18.2"
sqlalchemy2-stubs = "^0.0.2-alpha.21"
sqlalchemy-bigquery = "^1.4.3"
python-dateutil = "2.8.2"
requests = "^2.27.1"
MarkupSafe = "2.0.1"
psycopg2 = "^2.9.3"
google-auth = "^2.13.0"
google-cloud-bigquery = "2.34.2"
halo = "^0.0.31"
update-checker = "^0.18.0"
"ruamel.yaml" = "^0.17.21"
rudder-sdk-python = "^1.0.3"
duckdb-engine = "^0.10.0"
duckdb = "1.3.0"
yamllint = "^1.26.3"
click = ">=7.1.2"
GitPython = "^3.1.27"
databricks-sql-connector = "2.0.3"
dbt-snowflake = {version="^1.3.0", optional=true}
dbt-redshift = {version="^1.3.0", optional=true}
dbt-postgres = {version="^1.3.0", optional=true}
dbt-bigquery = {version="^1.3.0", optional=true}
dbt-metadata-client = {version="^0.1.0", optional=true}
grpcio = "^1.54"
pymysql = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest-mock = "^3.7.0"
pytest = "^7.1.1"
pre-commit = "^2.18.0"

[tool.poetry.extras]
dbt-snowflake = ["dbt-snowflake"]
dbt-redshift = ["dbt-redshift"]
dbt-postgres = ["dbt-postgres"]
dbt-bigquery = ["dbt-bigquery"]
dbt-cloud = ["dbt-metadata-client"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mf = 'metricflow.cli.main:cli'
datus-metricflow = 'datus_metricflow.cli:cli'

[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38']
force-exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
